LayoutTests/imported/w3c/web-platform-tests/css/css-anchor-position/position-area-basic-expected.txt


PASS Offsets for: none
FAIL Offsets for: span-all assert_equals: Check expected offsetWidth expected 400 but got 0
FAIL Offsets for: span-all span-all assert_equals: Check expected offsetWidth expected 400 but got 0
FAIL Offsets for: top left assert_equals: Check expected offsetWidth expected 100 but got 0
FAIL Offsets for: top center assert_equals: Check expected offsetLeft expected 100 but got 0
FAIL Offsets for: top right assert_equals: Check expected offsetLeft expected 250 but got 0
FAIL Offsets for: center left assert_equals: Check expected offsetTop expected 150 but got 0
FAIL Offsets for: center center assert_equals: Check expected offsetLeft expected 100 but got 0
FAIL Offsets for: center right assert_equals: Check expected offsetLeft expected 250 but got 0
FAIL Offsets for: bottom left assert_equals: Check expected offsetTop expected 225 but got 0
FAIL Offsets for: bottom center assert_equals: Check expected offsetLeft expected 100 but got 0
FAIL Offsets for: bottom right assert_equals: Check expected offsetLeft expected 250 but got 0
FAIL Offsets for: start start assert_equals: Check expected offsetWidth expected 100 but got 0
FAIL Offsets for: start center assert_equals: Check expected offsetLeft expected 100 but got 0
FAIL Offsets for: start end assert_equals: Check expected offsetLeft expected 250 but got 0
FAIL Offsets for: center start assert_equals: Check expected offsetTop expected 150 but got 0
FAIL Offsets for: center end assert_equals: Check expected offsetLeft expected 250 but got 0
FAIL Offsets for: end start assert_equals: Check expected offsetTop expected 225 but got 0
FAIL Offsets for: end center assert_equals: Check expected offsetLeft expected 100 but got 0
FAIL Offsets for: end end assert_equals: Check expected offsetLeft expected 250 but got 0
FAIL Offsets for: self-start self-start assert_equals: Check expected offsetWidth expected 100 but got 0
FAIL Offsets for: self-start center assert_equals: Check expected offsetLeft expected 100 but got 0
FAIL Offsets for: self-start self-end assert_equals: Check expected offsetLeft expected 250 but got 0
FAIL Offsets for: center self-start assert_equals: Check expected offsetTop expected 150 but got 0
FAIL Offsets for: center self-end assert_equals: Check expected offsetLeft expected 250 but got 0
FAIL Offsets for: self-end self-start assert_equals: Check expected offsetTop expected 225 but got 0
FAIL Offsets for: self-end center assert_equals: Check expected offsetLeft expected 100 but got 0
FAIL Offsets for: self-end self-end assert_equals: Check expected offsetLeft expected 250 but got 0
FAIL Offsets for: y-start x-start assert_equals: Check expected offsetWidth expected 100 but got 0
FAIL Offsets for: y-start center assert_equals: Check expected offsetLeft expected 100 but got 0
FAIL Offsets for: y-start x-end assert_equals: Check expected offsetLeft expected 250 but got 0
FAIL Offsets for: center x-start assert_equals: Check expected offsetTop expected 150 but got 0
FAIL Offsets for: center x-end assert_equals: Check expected offsetLeft expected 250 but got 0
FAIL Offsets for: y-end x-start assert_equals: Check expected offsetTop expected 225 but got 0
FAIL Offsets for: y-end center assert_equals: Check expected offsetLeft expected 100 but got 0
FAIL Offsets for: y-end x-end assert_equals: Check expected offsetLeft expected 250 but got 0
FAIL Offsets for: y-self-start x-self-start assert_equals: Check expected offsetWidth expected 100 but got 0
FAIL Offsets for: y-self-start center assert_equals: Check expected offsetLeft expected 100 but got 0
FAIL Offsets for: y-self-start x-self-end assert_equals: Check expected offsetLeft expected 250 but got 0
FAIL Offsets for: center x-self-start assert_equals: Check expected offsetTop expected 150 but got 0
FAIL Offsets for: center x-self-end assert_equals: Check expected offsetLeft expected 250 but got 0
FAIL Offsets for: y-self-end x-self-start assert_equals: Check expected offsetTop expected 225 but got 0
FAIL Offsets for: y-self-end center assert_equals: Check expected offsetLeft expected 100 but got 0
FAIL Offsets for: y-self-end x-self-end assert_equals: Check expected offsetLeft expected 250 but got 0
FAIL Offsets for: span-y-self-start span-x-self-end assert_equals: Check expected offsetLeft expected 100 but got 0
FAIL Offsets for: span-bottom span-all assert_equals: Check expected offsetTop expected 150 but got 0
PASS Offsets for: span-all top
